chore[fsst]: more *shared* benchmarks
