+++ encodings/fsst/Cargo.toml
@@ -30,27 +30,25 @@ vortex-session = { workspace = true }
 _test-harness = ["dep:rand", "vortex-array/_test-harness"]
 
 [dev-dependencies]
-aho-corasick = { workspace = true }
-daachorse = { workspace = true }
 divan = { workspace = true }
-jetscii = { workspace = true }
-memchr = { workspace = true }
 rand = { workspace = true }
-regex-automata = { workspace = true }
 rstest = { workspace = true }
 vortex-array = { workspace = true, features = ["_test-harness"] }
 
 [[bench]]
 name = "fsst_compress"
 harness = false
+required-features = ["_test-harness"]
 
 [[bench]]
 name = "fsst_contains"
 harness = false
+required-features = ["_test-harness"]
 
 [[bench]]
 name = "fsst_url_compare"
 harness = false
+required-features = ["_test-harness"]
 
 [[bench]]
 name = "chunked_dict_fsst_builder"
